HACKATHON Project
This Code was made by: Victor Garza, Roberto Jose & Diego Govea

This project was created in Java, we decided to make some JFrames with some functionality for showing the idea of the programm. This project uses OOP for the movement of data, it also uses a hash encryption method that helps the password be a bit more safe (We know isn't that safe but is something, we could add the "SALT" for making the hash unique for every user and we could also made a bcrypt method).